third_party/WebKit/LayoutTests/inspector-protocol/heap-profiler/heap-snapshot-with-active-dom-object-expected.txt

CONSOLE MESSAGE: line 21: Created 2 MediaQueryList eleements
Test that all ActiveDOMObjects with pending activities will get into one group in the heap snapshot. Bug 426809.

 Took heap snapshot
Parsed snapshot
FAIL: cannot find '(Pending activities group)'
